requirements.txt
aiogram>=3.6,<4.0
python-dotenv>=1.0.1
openai>=1.40.0
httpx>=0.27.0
